LayoutTests/fast/events/touch/gesture/touch-gesture-scroll-remove-node.html

<!DOCTYPE=html>
<html>
<head>
<script src="../../../js/resources/js-test-pre.js"></script>
<script src="resources/gesture-helpers.js"></script>
<style>
  #spacing {
    width: 200px;
    height: 120px;
  }
  #scroller {
    border: 1px solid #777;
    box-sizing: content-box;
    height: 300px;
    margin: 20px;
    overflow-x: hidden;
    overflow-y: auto;
    width: 100px
  }
  .spacer {
    box-sizing: border-box;
    display: block;
    overflow: hidden;
    visibility: hidden;
  }
  .list-item {
    line-height: 30px;
    text-align: center;
  }
</style>
</head>
<body onload="runTest();">
  <div id="spacing"></div>
  <div id="scroller">
    <div id = "top-spacer" class="spacer"></div>

    <div id = "bottom-spacer" class="spacer"></div>
  </div>
<script>
var cache = {};

/* One more that the maximum that can be fully fit on screen. */
var maxVisibleItems = 11;

var listSize = 300;

var itemHeight = 30;

function $(name) {
  return document.getElementById(name);
}

function onScroll(e) {
  redraw();
}

function redraw() {
  // Create new items and add to cache.
  var top = $('scroller').scrollTop;
  var first = Math.floor(top / 30);
  if (first > listSize - maxVisibleItems)
    first = listSize - maxVisibleItems;
  var newItems = [];
  for (var i = 0; i < maxVisibleItems; i++) {
    var index = first + i;
    if (cache[index] == undefined) {
      var element = document.createElement('div');
      element.className = 'list-item';
      element.textContent = String(index+1);
      element.data = {};
      element.data.index = index;
      cache[index] = element;
      newItems.push(element);
    }
  }
  // Remove items that are not visible.
  for (var i = 0; i < first; i++) {
    if (cache[i])
      delete cache[i];
  }
  for (var i = first + maxVisibleItems; i < listSize; i++) {
    if (cache[i])
      delete cache[i];
  }
  // Pad spacer elements to preserve height of scrollable list.
  $('top-spacer').style.height = (first * itemHeight) + 'px';
  $('bottom-spacer').style.height =
      ((listSize - maxVisibleItems - first) * 30) + 'px';

  // Remove off-screen elements.
  var anchor = $('bottom-spacer');
  var parent = $('scroller');
  for (var item = $('top-spacer').nextSibling; item != anchor; ) {
    if (!item.classList || !item.classList.contains('list-item')) {
      item = item.nextSibling;
      continue;
    }
    if (!cache[item.data.index]) {
      var next = item.nextSibling;
      scroller.removeChild(item);
      item = next;
    } else {
      item = item.nextSibling;
    }
  }

  // Insert new on-screen elements.
  var referenceElement = $('top-spacer').nextSibling;
  for (var i = 0; i < newItems.length; i++) {
    var index = newItems[i].data.index;
    while (referenceElement != anchor && (!referenceElement.classList ||
           !referenceElement.classList.contains('list-item') ||
           referenceElement.data.index < index)) {
      referenceElement = referenceElement.nextSibling;
    }
    parent.insertBefore(newItems[i], referenceElement);
  }
}

$('scroller').addEventListener('scroll', onScroll);
redraw();

if (window.testRunner)
    testRunner.waitUntilDone();

var scrollCount = 0;
var wheelCount = 0;
var nextStep;
var scrollIndex = 2;
var startPositionX;
var startPositionY;
var scrollAmount;

function firstStep() {
    debug("begin scroll");
    var item = cache[scrollIndex];
    startPositionX = item.offsetLeft + 5;
    startPositionY = item.offsetTop + 5;
    eventSender.gestureScrollBegin(startPositionX, startPositionY);

    var aboveScroller = scroller.offsetTop - 30;
    scrollAmount = startPositionY - aboveScroller;
    debug("scroll up");

    nextStep = secondStep;

    eventSender.gestureScrollUpdate(0, -scrollAmount);
    shouldBe('scroller.scrollTop', '' + scrollAmount);
    shouldBe('scroller.scrollLeft', '0');
    shouldBe('wheelCount', '1');
}

function secondStep() {
    shouldBe('scrollCount', '1');
    shouldBe('typeof cache[' + scrollIndex + ']', 'typeof undefined');
    debug("scroll down");

    nextStep = thirdStep;

    eventSender.gestureScrollUpdate(0, scrollAmount);
    shouldBe('scroller.scrollTop', '0');
    shouldBe('scroller.scrollLeft', '0');
    shouldBe('wheelCount', '2');
}

function thirdStep() {
    shouldBe('scrollCount', '2');
    testRunner.notifyDone();
}

function runTest()
{
    var scroller = $('scroller');
    scroller.addEventListener("scroll", function() {
        ++scrollCount;
        setTimeout(function() {
            nextStep();
        }, 0);
    });
    window.addEventListener("mousewheel", function() {
        ++wheelCount;
    });

    if (window.eventSender && window.testRunner) {
        description('This tests gesture event scrolling of a div where the element the scrolling starts on is removed from an onscroll event handler.');
        if (checkTestDependencies()) {
            firstStep();
        } else {
            exitIfNecessary();
        }
    } else {
        debug("This test requires DumpRenderTree.");
    }
}
</script>
</body>
</html>
