LayoutTests/fast/events/touch/gesture/touch-gesture-scroll-remove-node-expected.txt

This tests gesture event scrolling of a div where the element the scrolling starts on is removed from an onscroll event handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


begin scroll
scroll up
PASS scroller.scrollTop is 96
PASS scroller.scrollLeft is 0
PASS wheelCount is 1
PASS scrollCount is 1
PASS typeof cache[2] is typeof undefined
scroll down
PASS scroller.scrollTop is 0
PASS scroller.scrollLeft is 0
PASS wheelCount is 2
PASS scrollCount is 2
1
2
3
4
5
6
7
8
9
10
11
